+++ lib/src/theme/base_theme.dart
@@ -16,33 +16,39 @@ class BaseTheme {
     Brightness brightness = Brightness.light,
   }) {
     return (Brightness.light == brightness ? ThemeData.light() : ThemeData.dark()).copyWith(
-        colorScheme: ColorScheme.fromSeed(
-          seedColor: color,
-          primary: color,
-          brightness: brightness,
-          surfaceTint: BaseColors.cTransparent,
-        ),
-        iconTheme: const IconThemeData(
-          size: BaseDimens.dIconSize,
-          fill: 0.0,
-          weight: 400.0,
-          grade: 0.0,
-          opticalSize: BaseDimens.dIconBtnHeight,
-          color: BaseColors.cWhite,
-          // opacity: 0.8,
-        ),
-        floatingActionButtonTheme: FloatingActionButtonThemeData(
-          foregroundColor: BaseColors.cWhite,
-          backgroundColor: color,
-          shape: const CircleBorder(),
-        ),
-        appBarTheme: defaultAppBarTheme,
-        tabBarTheme: const TabBarTheme(
-          dividerHeight: 0,
-          indicatorSize: TabBarIndicatorSize.label,
-          indicatorColor: BaseColors.cFontWhite,
-          labelPadding: EdgeInsets.zero,
-        ));
+      colorScheme: ColorScheme.fromSeed(
+        seedColor: color,
+        primary: color,
+        brightness: brightness,
+        surfaceTint: BaseColors.cTransparent,
+      ),
+      iconTheme: const IconThemeData(
+        size: BaseDimens.dIconSize,
+        fill: 0.0,
+        weight: 400.0,
+        grade: 0.0,
+        opticalSize: BaseDimens.dIconBtnHeight,
+        color: BaseColors.cWhite,
+        // opacity: 0.8,
+      ),
+      floatingActionButtonTheme: FloatingActionButtonThemeData(
+        foregroundColor: BaseColors.cWhite,
+        backgroundColor: color,
+        shape: const CircleBorder(),
+      ),
+      appBarTheme: defaultAppBarTheme,
+      tabBarTheme: const TabBarTheme(
+        dividerHeight: 0,
+        indicatorSize: TabBarIndicatorSize.label,
+        indicatorColor: BaseColors.cFontWhite,
+        labelPadding: EdgeInsets.zero,
+      ),
+      // 自定义
+      textTheme: defaultTextTheme,
+      inputDecorationTheme: const InputDecorationTheme(
+        hintStyle: TextStyle(color: BaseColors.cFontGrayLight, height: 1.2),
+      ),
+    );
   }
 
   static AppBarTheme createAppBarTheme({
@@ -52,16 +58,42 @@ class BaseTheme {
     double? elevation,
     SystemUiOverlayStyle overlayStyle = SystemUiOverlayStyle.light,
     Brightness brightness = Brightness.light,
+    bool? centerTitle,
   }) {
     return (brightness == Brightness.light ? ThemeData.light() : ThemeData.dark()).appBarTheme.copyWith(
           iconTheme: IconThemeData(
             color: iconColor,
             size: BaseDimens.dIconSize,
           ),
+          centerTitle: centerTitle,
           elevation: elevation,
           backgroundColor: bgColor,
           titleTextStyle: titleTextStyle ?? Typography.dense2014.titleLarge,
           systemOverlayStyle: overlayStyle,
         );
   }
+
+  static TextTheme defaultTextTheme = const TextTheme(
+    bodyLarge: TextStyle(
+      color: BaseColors.cFontBlackLight,
+      fontSize: BaseDimens.dFontSizeLarge,
+      overflow: TextOverflow.ellipsis,
+      height: 1.3,
+      decoration: TextDecoration.none,
+    ),
+    bodyMedium: TextStyle(
+      color: BaseColors.cFontBlack,
+      fontSize: BaseDimens.dFontSizeNormal,
+      overflow: TextOverflow.ellipsis,
+      height: 1.3,
+      decoration: TextDecoration.none,
+    ),
+    bodySmall: TextStyle(
+      color: BaseColors.cFontGrayLight,
+      fontSize: BaseDimens.dFontSizeSmall,
+      overflow: TextOverflow.ellipsis,
+      height: 1.3,
+      decoration: TextDecoration.none,
+    ),
+  );
 }